Move mods to packages of mod category
