+++ src/main/java/com/peasenet/mods/movement/ModClimb.java
@@ -18,28 +18,18 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.peasenet.mods;
+package com.peasenet.mods.movement;
 
-import com.peasenet.main.GavinsModClient;
-import net.minecraft.item.BlockItem;
+import com.peasenet.mods.Mod;
+import com.peasenet.mods.Type;
 
 /**
  * @author gt3ch1
  * @version 6/14/2022
- * A mod that allows the player to place blocks really fast.
+ * A mod that allows the player to climb up walls despite a lack of ladders.
  */
-public class ModFastPlace extends Mod {
-    public ModFastPlace() {
-        super(Type.FAST_PLACE);
-    }
-
-    @Override
-    public void onTick() {
-        super.onTick();
-        if (GavinsModClient.getPlayer() != null) {
-            var item = GavinsModClient.getPlayer().getMainHandStack().getItem() instanceof BlockItem;
-            if (isEnabled && item)
-                getClient().setItemUseCooldown(0);
-        }
+public class ModClimb extends Mod {
+    public ModClimb() {
+        super(Type.CLIMB);
     }
 }
