+++ src/main/java/com/peasenet/mods/movement/ModFastMine.java
@@ -18,15 +18,18 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.peasenet.mods;
+package com.peasenet.mods.movement;
+
+import com.peasenet.mods.Mod;
+import com.peasenet.mods.Type;
 
 /**
  * @author gt3ch1
  * @version 6/14/2022
- * A mod that allows the client to see boxes around mobs.
+ * A mod that allows the player to mine very quickly.
  */
-public class ModMobEsp extends Mod {
-    public ModMobEsp() {
-        super(Type.MOB_ESP);
+public class ModFastMine extends Mod {
+    public ModFastMine() {
+        super(Type.FAST_MINE);
     }
 }
