+++ src/main/java/com/peasenet/mods/movement/ModNoFall.java
@@ -18,15 +18,18 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.peasenet.mods;
+package com.peasenet.mods.movement;
+
+import com.peasenet.mods.Mod;
+import com.peasenet.mods.Type;
 
 /**
  * @author gt3ch1
  * @version 6/14/2022
- * A mod that allows the player to see an ESP to other players.
+ * A mod that prevents the player from taking fall damage.
  */
-public class ModEntityPlayerEsp extends Mod {
-    public ModEntityPlayerEsp() {
-        super(Type.ENTITY_PLAYER_ESP);
+public class ModNoFall extends Mod {
+    public ModNoFall() {
+        super(Type.NO_FALL);
     }
 }
