+++ src/main/java/com/peasenet/mods/render/ModAntiHurt.java
@@ -18,29 +18,18 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.peasenet.mods;
+package com.peasenet.mods.render;
 
-import com.peasenet.main.GavinsMod;
-import com.peasenet.main.GavinsModClient;
-import com.peasenet.util.KeyBindUtils;
+import com.peasenet.mods.Mod;
+import com.peasenet.mods.Type;
 
 /**
  * @author gt3ch1
  * @version 6/14/2022
- * A mod that allows the player to see a menu of all the mods by pressing the keybind.
+ * A mod that disables the view bob when the player is damaged.
  */
-public class ModGui extends Mod {
-    public ModGui() {
-        super(Type.MOD_GUI, KeyBindUtils.registerKeyBindForType(Type.MOD_GUI));
-    }
-
-    @Override
-    public void onEnable() {
-        GavinsModClient.getMinecraftClient().setScreen(GavinsMod.gui);
-    }
-
-    @Override
-    public void onDisable() {
-
+public class ModAntiHurt extends Mod {
+    public ModAntiHurt() {
+        super(Type.ANTI_HURT);
     }
 }
